refactor: :truck: move tests to test_check
